feat: gogo battery dying
